% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{compute_weighted_hazard}
\alias{compute_weighted_hazard}
\title{Calculate Observed Natural Course Risk}
\usage{
compute_weighted_hazard(
  prob_censor,
  data,
  id,
  censor_varname,
  time_points,
  time_name,
  outcome_name,
  competing_varname,
  competing_fit,
  total_effect
)
}
\arguments{
\item{prob_censor}{a vector of numerics specifying the estimated probability of censoring for each individual.}

\item{data}{a data frame containing the observed data.}

\item{id}{a character string indicating the ID variable name in \code{data}.}

\item{censor_varname}{a character string indicating the censor variable name in \code{data}.}

\item{time_points}{a numeric that indicates the total number of time points.}

\item{time_name}{a character string indicating the time variable name in \code{data}.}

\item{outcome_name}{a character string indicating the outcome variable name in \code{data}.}

\item{competing_varname}{a character string indicating the competing event variable name in \code{data}.}

\item{competing_fit}{a fitted model for the competing event model.}

\item{total_effect}{a logical indicating whether to treat competing event as censoring or total effect.}
}
\value{
A list with the first entry as a vector of the mean observed risk.
Its second entry is a vector of mean observed survival. Its third entry is a vector of inverse probability weight.
}
\description{
This functions calculates the inverse probability weighted observed natural course risk.
}
