% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICE.R
\name{construct_interv_value}
\alias{construct_interv_value}
\title{Construct intervened values for specified intervention time points}
\usage{
construct_interv_value(data, timevar, int_value, int_time, int_var)
}
\arguments{
\item{data}{a dataframe containing the observed data.}

\item{timevar}{a character string specifying the time index column.}

\item{int_value}{a list containing the intervened values for each intervention on all time points.}

\item{int_time}{a list containing the intervention time points for each intervention.}

\item{int_var}{a list containing the intervention variable names for each intervention.}
}
\value{
the intervened values on specified intervention time points.
}
\description{
Construct intervened values for specified intervention time points
}
\keyword{internal}
