% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICE.R
\name{get_model_formula}
\alias{get_model_formula}
\title{Get the model formula from keyword arguments}
\usage{
get_model_formula(interv_list, arg_interv, model_interv, ninterv, arg_str)
}
\arguments{
\item{interv_list}{the list of names of interventions.}

\item{arg_interv}{the list of arguments containing the target keyword argument.}

\item{model_interv}{the list of model inputs related to the target keyword argument.}

\item{ninterv}{the number of interventions.}

\item{arg_str}{the argument keyword.}
}
\value{
a list containing the model inputs corresponding to the interventions.
}
\description{
Get the model formula from keyword arguments
}
\keyword{internal}
