% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_rmse}
\alias{get_rmse}
\title{Get Root Mean Square Error from the Model Object}
\usage{
get_rmse(fit)
}
\arguments{
\item{fit}{a glm model object.}
}
\value{
the root mean square error (RMSE) values of the fitted model.
}
\description{
Get Root Mean Square Error from the Model Object
}
\keyword{internal}
