% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_stderr}
\alias{get_stderr}
\title{Get Standard Deviation from Model Object}
\usage{
get_stderr(fit)
}
\arguments{
\item{fit}{a glm model object.}
}
\value{
the standard errors of the coefficients of the fitted model.
}
\description{
Get Standard Deviation from Model Object
}
\keyword{internal}
