% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_summary}
\alias{get_summary}
\title{Get Summary Table from Model Object}
\usage{
get_summary(fit)
}
\arguments{
\item{fit}{a glm model object.}
}
\value{
the summary table of the fitted model.
}
\description{
Get Summary Table from Model Object
}
\keyword{internal}
