% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_vcov}
\alias{get_vcov}
\title{Get Variance-Covariance Matrix from Model Object}
\usage{
get_vcov(fit)
}
\arguments{
\item{fit}{a glm model object.}
}
\value{
the variance-covariance matrices of the parameters of the fitted model.
}
\description{
Get Variance-Covariance Matrix from Model Object
}
\keyword{internal}
