% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{grace_period}
\alias{grace_period}
\title{}
\usage{
grace_period(
  type,
  nperiod,
  condition,
  data = interv_data,
  id = idvar,
  time_name = time0var,
  outcome_name = outcomevar
)
}
\arguments{
\item{type}{a character string specifying the type of grace period strategy. "Uniform" for uniform grace period, and "natural" for natural grace period.}

\item{nperiod}{a numeric specifying the length of grace period.}

\item{condition}{a character string specifying the logical expression, upon which is met, the treatment is initiated in \code{n} time units, where \code{n} is specified by \code{nperiod}.}

\item{data}{a data frame containing the observed data.}

\item{id}{a character string indicating the ID variable name in \code{data}.}

\item{time_name}{a character string indicating the time variable name in \code{data}.}

\item{outcome_name}{a character string indicating the outcome variable name in \code{data}.}
}
\value{
a vector containing the intervened value in the same size as the number of rows in \code{data}.
}
\description{

}
\details{

}
\examples{
data <- readRDS("test_data_competing.rds")
grace_period <- grace_period(type = "uniform", nperiod = 2, condition = "L1 == 0", data = data,
                             id = "id", time_name = "t0", outcome_name = "Y")
grace_period
}
