% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ice.R
\name{interv_boot}
\alias{interv_boot}
\title{Get the intervened values on bootstrap sample}
\usage{
interv_boot(interv, idx)
}
\arguments{
\item{interv}{a list containing the intervened values corresponding to each intervention.}

\item{idx}{a vector of indices to be selected for bootstrap sample.}
}
\value{
a list containing the intervened values for bootstrap sample.
}
\description{
Get the intervened values on bootstrap sample
}
\keyword{internal}
