% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICE.R
\name{match_boot_values}
\alias{match_boot_values}
\title{Match bootstrap values to aggregate data frame}
\usage{
match_boot_values(risk_df, cv_df, col)
}
\description{
Match bootstrap values to aggregate data frame
}
\keyword{internal}
