% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{natural_course_ipweighted}
\alias{natural_course_ipweighted}
\title{IPW Estimator for Natural Course Risk}
\usage{
natural_course_ipweighted(
  data,
  id,
  censor_varname,
  K,
  time0,
  outcome_varname,
  covar,
  competing_varname,
  competing_fit,
  total_effect
)
}
\arguments{
\item{data}{a data frame containing the observed data.}

\item{id}{a character string indicating the ID variable name in \code{data}.}

\item{censor_varname}{a character string indicating the censoring event variable name in \code{data}.}

\item{K}{a numeric that indicates the total number of time points.}

\item{time0}{a character string indicating the time variable name in \code{data}.}

\item{outcome_varname}{a character string indicating the outcome variable name in \code{data}.}

\item{covar}{a vector of character strings specifying the covariates used
in the probability of censoring model for the inverse probability weighted estimate of the natural course risk.}

\item{competing_varname}{a character string indicating the competing event variable name in \code{data}.}

\item{competing_fit}{a fitted model for the competing event model.}

\item{total_effect}{a logical indicating whether to treat competing event as censoring or total effect.}
}
\value{
A list containing the inverse probability weighted estimates of the natural course risk and
a model object for the probability of censoring used in the inverse probability weighted estimate of the natural course risk.
}
\description{
This is the internal helper function for the ICE estimator for the inverse probability weighted estimate of the natural course risk
}
\keyword{internal}
