% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{product}
\alias{product}
\title{Calculate product of multiple columns in a data frame}
\usage{
product(data, columns)
}
\arguments{
\item{data}{a data frame containing the observed data.}

\item{columns}{a vector of character strings specifying the variables in \code{data} to be used to calculate the product.}
}
\value{
a vector of the product of the variables specified in \code{columns}.
}
\description{
This function is an internal function that helps calculate the product of multiple columns in the data frame.
}
\keyword{internal}
