% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICE.R
\name{split_args}
\alias{split_args}
\title{Split Keyword Arguments}
\usage{
split_args(argument, target_string)
}
\arguments{
\item{argument}{the keyword argument list.}

\item{target_string}{a string specifying the target keyword argument pattern.}
}
\value{
a list containing the information of the targeted keyword argument.
}
\description{
Split Keyword Arguments
}
\keyword{internal}
