% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{static}
\alias{static}
\title{Static}
\usage{
static(value, data)
}
\arguments{
\item{value}{a number specifying the intervention value. 0 for never treat.}

\item{data}{a data frame containing the observed data.}
}
\value{
a vector containing the intervened values of the same size as the number of rows in \code{data}.
}
\description{
This function specifies the static intervention, either treat with a constant value or never treat, 
on the treatment variable in \code{data}.
}
\examples{
data <- gfoRmulaICE::compData
always_treat <- static(value = 1, data = data)
}
