% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{summary_table}
\alias{summary_table}
\title{Summary Table for ICE Estimator Objects}
\usage{
summary_table(...)
}
\arguments{
\item{...}{the ICE estimator objects.}
}
\value{
a data frame containing the summary table for all specified ICE estimator objects.
}
\description{
This function returns a summary table for ICE estimator objects. (This function is going to be converted to the S3 method in R so it is named "summary_table" for now.
After conversion, users could use the base function summary().)
}
\examples{

fit_classical_pool <- ice(
data = data,
K = 4,
id = "id",
time_name = "t0",
outcome_name = "Y",
censor_name = "C",
competing_name = "D",
estimator = pool(hazard = F),
comp_effect = 0,
outcome_model = Y ~ L1 + L2 + A1 + A2,
censor_model = C ~ L1 + L2 + A1 + A2,
ref_idx = 0,
int_descript = c("Static Intervention"),
intervention1.A1 = list(static(3)),
intervention1.A2 = list(static(1))
)

fit_hazard_pool <- ice(
data = data,
K = 4,
id = "id",
time_name = "t0",
outcome_name = "Y",
censor_name = "C",
competing_name = "D",
estimator = pool(hazard = T),
comp_effect = 0,
outcome_model = Y ~ L1 + L2 + A1 + A2,
censor_model = C ~ L1 + L2 + A1 + A2,
competing_model = D ~ L1 + L2 + A1 + A2,
ref_idx = 0,
int_descript = c("Static Intervention"),
intervention1.A1 = list(static(3)),
intervention1.A2 = list(static(1))
)

summary_table(fit_classical_pool, fit_hazard_pool)
}
