% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{threshold}
\alias{threshold}
\title{Threshold Intervention}
\usage{
threshold(
  lower_bound,
  upper_bound,
  var = threshold_treatment,
  data = interv_data
)
}
\arguments{
\item{lower_bound}{a numeric specifying the lower bound of the threshold.}

\item{upper_bound}{a numeric specifying the upper bound of the threshold.}

\item{var}{a character string specifying the treatment variable for the intervention.}

\item{data}{a data frame containing the observed data.}
}
\value{
the intervened values on the intervention variable.
}
\description{
This function implements the threshold intervention.
At each time, if an individual's treatment value is within the user-specified range inclusively,
then follow the natural value of the treatment. Otherwise, if the treatment value is below the
lower bound specified by the user, then set to the lower bound of the range,
and if the treatment value is above the upper bound specified by the user, then set to the upper bound of the range.
For more details, please see Young et al 2014.
}
\examples{
data <- readRDS("test_data_competing.rds")
threshold_treat <- threshold(threshold = -3, var = "A", data = data)
threshold_treat
}
\references{
Young JG, Herńan MA, Robins JM. Identification, estimation and approximation of risk under interventions that depend on the natural value of treatment using observational data. Epidemiologic Methods. 2014;3(1):1-19.
}
