% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{uniform_sample}
\alias{uniform_sample}
\title{Uniform Sampling}
\usage{
uniform_sample(r, duration)
}
\arguments{
\item{r}{a vector of numerics specifying the time until initiate the treatment.}

\item{duration}{a numeric specifying the duration of grace period.}
}
\value{
the randomly generated treatment.
}
\description{
This function is an internal function that helps with the implementation of uniform
grace period intervention function.
}
\keyword{internal}
