% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ice.R
\name{bootstrap_ice}
\alias{bootstrap_ice}
\title{Bootstrap for ICE estimator}
\usage{
bootstrap_ice(
  f,
  K,
  nboot,
  sig_level,
  parallel,
  ncores,
  ref_description,
  ref_intervention_varname,
  ref_total_effect,
  total_effect,
  natural_course_type,
  ref_intervention,
  interventions,
  intervention_varnames,
  intervention_description,
  data,
  id,
  set_seed,
  ...
)
}
\arguments{
\item{f}{a function specifying which ICE estimator to use for bootstrap.}

\item{K}{a numerical value that indicates the total number of time points.}

\item{nboot}{a numeric indicating number of bootstrap samples.}

\item{sig_level}{a numeric indicating the significance level to be used for confidence interval.}

\item{parallel}{a logical indicating whether to parallelize the bootstrap process.}

\item{ncores}{a numeric indicating the number of CPU cores to be used in parallel computing.}

\item{ref_description}{a string specifying a description for the reference intervention.}

\item{ref_intervention_varname}{a string specifying the intervention variable used in the reference intervention.}

\item{ref_total_effect}{a logical indicating how the competing event is handled for the reference intervention.
TRUE for total effect. FALSE for direct effect.}

\item{total_effect}{a logical indicating how the competing event is handled for the defined intervention.
TRUE for total effect. FALSE for direct effect.}

\item{natural_course_type}{a character string indicating how the natural course strategy is computed for the defined intervention.
"ipw" for IPW natural course risk estimation.}

\item{ref_intervention}{a list of function specifying the intervention to be used as reference.
for the inverse probability weighted estimate of the natural course risk in the reference intervention.}

\item{interventions}{a list of functions defining the interventions to be used in this bootstrap.}

\item{intervention_varnames}{a list of character strings corresponding the interventions to be used in this bootstrap.}

\item{intervention_description}{a character string specifying a description for the define intervention in this bootstrap.}

\item{data}{a data frame containing the observed data in long format.}

\item{id}{a character string indicating the ID variable name in \code{data}.}

\item{set_seed}{a numeric indicating the starting seed for bootstrap.}

\item{...}{any keyword arguments to be passed in \code{f}.}
}
\value{
A list containing the following components:
\item{ice_se}{The standard error for the ICE risk of the defined intervention.}
\item{ref_se}{The standard error for the ICE risk of the reference intervention.}
\item{rr_se}{The standard error for the ICE risk of the risk ratio.}
\item{rd_se}{The standard error for the ICE risk of the risk difference.}
\item{ice_cv}{The empirical quantile for the ICE risk of the defined intervention.}
\item{ref_cv}{The empirical quantile for the ICE risk of the reference intervention.}
\item{rr_cv}{The empirical quantile for the ICE risk of the risk ratio.}
\item{rd_cv}{The empirical quantile for the ICE risk of the risk difference.}
\item{boot_data}{A list containing all the bootstrapped data samples.}
\item{boot_models}{A list containing the fitted models for the outcome, the treatment (if applicable), and the competing event (if applicable) on the bootstrapped samples.}
\item{boot_summary}{A list containing the summary of the fitted models on the bootstrapped samples.}
\item{boot_stderr}{A list containing the standard errors of the coefficients of the fitted models on the bootstrapped samples.}
\item{boot_vcov}{A list containing the variance-covariance matrices of the parameters of the fitted models on the bootstrapped samples.}
\item{boot_rmse}{A list containing the root mean square error (RMSE) values of the fitted models on the bootstrapped samples.}
}
\description{
This function estimates the variance of the point estimates for all user-defined interventions from
the ICE estimator object. The variance is calculated by bootstrapping and is used to calculate the confidence
interval for each point estimate of each user-defined intervention specified in the ICE estimator object.
}
