% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dynamic}
\alias{dynamic}
\title{Dynamic Intervention Function}
\usage{
dynamic(
  condition,
  strategy_before,
  strategy_after,
  absorb = FALSE,
  id = id_var,
  time = time0var,
  data = interv_data
)
}
\arguments{
\item{condition}{a character string that specifies a logical expression, upon which is met, the strategy specified in \code{strategy_after} is followed.}

\item{strategy_before}{a function or vector of intervened values in the same length as the number of rows in the observed data \code{data} that specifies the strategy followed after the condition in \code{condition} is met.}

\item{strategy_after}{a function or vector of intervened values in the same length as the number of rows in the observed data \code{data} that specifies the strategy followed before the condition in \code{condition} is met.}

\item{absorb}{a logical indicating whether the strategy specified in \code{strategy_after} is absorbing upon the first time when the condition specified in \code{condition} is met.}

\item{id}{a character string indicating the ID variable name in \code{data}.}

\item{time}{a character string indicating the time variable name in \code{data}.}

\item{data}{a data frame containing the observed data.}
}
\value{
a vector containing the intervened value in the same size as the number of rows in \code{data}.
}
\description{
This function implements the dynamic intervention strategy. Dynamic intervention strategy refers to the class of treatment mechanisms that depend on the covariate values. 
This function provides an example of dynamic intervention strategy. The specific mechanism is described below:
\itemize{
\item Follow the strategy specified in \code{strategy_before} until a user-defined condition dependent on covariate values is met. 
\item Upon the user-defined condition is met, the strategy specified in \code{strategy_after} is initiated.
}
}
\examples{
data <- readRDS("test_data_competing.rds")
# Dynamic intervention example 1: 
# Treat when L1 = 0, and not treat otherwise.
dynamic <- dynamic(condition = "L1 == 0", strategy_before = static(0), strategy_after = static(1), 
absorb = FALSE, id = "id", time_name = "t0", data = data)
dynamic
}
