% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dynamic}
\alias{dynamic}
\title{Dynamic Intervention Function}
\usage{
dynamic(type, var, direction, value, id = id_var, data = interv_data)
}
\arguments{
\item{type}{a character string specifying the type of dynamic intervention. Possible inputs are \code{compare} and \code{absorbing}.}

\item{var}{a character string specifying the intervention variable name in \code{data}.}

\item{direction}{a character string specifying the comparison of \code{var} and \code{value}.
Possible inputs are \code{>, <, >=, <=, =, !=}.}

\item{value}{a numeric specifying the comparison value on \code{var}.}

\item{id}{a character string specifying the id variable in \code{data}.}

\item{data}{a data frame containing the observed data.}
}
\value{
a vector containing the intervened value in the same size as the number of rows in \code{data}.
}
\description{
This function implements the dynamic intervention strategy. Dynamic intervention strategy refers to the class of treatment mechanisms that depend on the covariate values. This function provides an example of dynamic intervention strategy. The specific mechanism is described as following:
\itemize{
\item For \code{type} being \code{compare}, the treatment is determined based on the condition of \code{direction} and \code{value}.
For example, if \code{direction} is >, then treat if \code{var} > \code{value}, and not treat otherwise.
\item For \code{type} being \code{absorbing}, the initiation of treatment is based on the same mechanism as \code{type} being \code{compare}, and
adopts always treat once the treatment is initiated for each individual.
}
Possible \code{direction} includes \code{>, <, >=, <=, =, !=}.
}
\examples{
data <- readRDS("test_data_competing.rds")
dynamic <- dynamic(type = "absorbing", var = "L1", value = 0, direction = ">", id = "id", data = data)
dynamic
}
