% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICE.R
\name{ice}
\alias{ice}
\title{Singly Robust and Doubly Robust Iterative Conditional Expectation (ICE) Estimator for the Specified Intervention}
\usage{
ice(
  data,
  time_points,
  id,
  time_name,
  outcome_name,
  censor_name = NULL,
  compevent_name = NULL,
  comp_effect = 0,
  outcome_model,
  censor_model = NULL,
  competing_model = NULL,
  ref_idx = 0,
  estimator,
  int_descript,
  ci_method = "percentile",
  nsamples = 0,
  seed = 1,
  significance_level = 0.05,
  parallel = F,
  ncores = 2,
  ...
)
}
\arguments{
\item{data}{a data frame containing the observed data in long format.}

\item{time_points}{a numerical value that indicates the total number of time points.}

\item{id}{a character string indicating the ID variable name in \code{data}.}

\item{time_name}{a character string indicating the time variable name in \code{data}.}

\item{outcome_name}{a character string indicating the outcome variable name in \code{data}.}

\item{censor_name}{a character string indicating the censor variable name in \code{data}. Default is \code{NULL}.}

\item{compevent_name}{a character string indicating the competing variable name in \code{data}. Default is \code{NULL}.}

\item{comp_effect}{a numeric specifying how the competing event is handled for all the specified interventions. Default is 0.
\code{0} outputs direct effect for all the specified interventions.
\code{1} outputs total effect for all the specified interventions.}

\item{outcome_model}{a formula specifying the model statement for the outcome model.}

\item{censor_model}{a formula specifying the model statement for the censoring model for IP weighted natural course risk. Default is \code{NULL}.}

\item{competing_model}{a formula specifying the model statement for the competing model for hazard-based ICE estimator. Default is \code{NULL}.}

\item{ref_idx}{a numerical indicating which intervention to be used as the reference to calculate the risk ratio and risk difference. Default is 0.
0 refers to using the natural course as the reference intervention.
Any other numbers refer to the corresponding intervention that users specify in the keywords argument.}

\item{estimator}{a function to specifying which ICE estimator to use for the estimation. Possible inputs are:
\itemize{
\item{Classical Pooling over Treatment History ICE Estimator: }{\code{pool(hazard = F)}}
\item{Hazard Based Pooling over Treatment History ICE Estimator: }{\code{pool(hazard = T)}}
\item{Classical Stratifying Treatment History ICE Estimator: }{\code{strat(hazard = F)}}
\item{Hazard Based Stratifying Treatment History ICE Estimator: }{\code{strat(hazard = T)}}
\item{Doubly Robust Weighted ICE Estimator: }{\cr
\code{weight(treat_model)}
where \code{treat_model} is a list specifying the treatment model.}
}}

\item{int_descript}{a vector of strings containing description for each specified intervention.}

\item{ci_method}{a character string specifying the method used for calculating the confidence interval, if \code{nsamples} is larger than 0.
Either "percentile" or "normal." Default is "percentile."}

\item{nsamples}{a numeric indicating number of bootstrap samples. If a number larger than 0 is specified, bootstrap samples will be used for
standard error estimate and confidence interval. Default is 0.}

\item{seed}{a numeric indicating the starting seed for bootstrap. Default is 1.}

\item{significance_level}{a numeric indicating the significance level to be used for confidence interval. Default is 0.05.}

\item{parallel}{a logical indicating whether to parallelize the bootstrap process. TRUE for using parallel computing. FALSE for not using parallel computing.
Default is FALSE.}

\item{ncores}{a numeric indicating the number of CPU cores to be used in parallel computing. Default is 2.}

\item{...}{keywords arguments for intervention inputs, optional outcome models for stratified ICE, and optional competing models for stratified ICE.
The keyword argument for interventions should follow the below naming convention:
\itemize{
\item{Each intervention is specified using the keyword argument name with \emph{intervention} prefix.}
\item{Use \emph{i} after \emph{intervention} prefix in keyword argument name to represent the ith intervention strategy.}
\item{Use \emph{.} followed with \emph{treatment variable name} after \emph{interventioni} in keyword argument name to represent the treatment name within the ith intervention strategy.}
}
The input for each keyword argument must be a list encompassing two elements, which are:
a vector of intervened values and an optional vector of time points on which the corresponding intervention is applied.
If the second element is not specified, then the defined intervention will be applied to all time points.
A sample intervention input with two treatments of names A1 and A2 and two intervention strategies look like: \cr
\code{intervention1.A1 = list(static(1))} \cr
\code{intervention1.A2 = list(dynamic("compare", "L1", ">", 0))} \cr
\code{intervention2.A1 = list(static(0))} \cr
\code{intervention2.A2 = list(dynamic("absorbing", "L1", ">", 1))} \cr
A sample intervention input with one treatment of name A and two intervention strategies look like: \cr
\code{intervention1.A1 = list(static(1))} \cr
\code{intervention2.A1 = list(static(0))} \cr
For the above two intervention inputs, since there is no second element input, each specified intervention is applied on all time points.
If one wants to specify the custom time points on which each intervention is applied, a sample intervention input looks like:
\code{intervention1.A1 = list(static(1), 1:2)} \cr
\code{intervention1.A2 = list(static(0), 3:5)} \cr
where the intervention on A1 within intervention strategy 1 is applied on time 1 and 2,
and the intervention on A2 within intervention strategy 1 is applied on time 3 to 5.
If there is no intervention specified, only the natural course risk will be returned. \cr
To specify different outcome model for different intervention, please follow the keyword argument input convention below:

\itemize{
\item{Each outcome model is specified using keyword argument name starting with \emph{outcomeModel} prefix.}
\item{Use \emph{.n} \emph{outcomeModel} prefix in keyword argument name to specify which intervention being applied to,
where \emph{n} represents the \emph{n}th intervention.}
}

To specify different competing model for different intervention, please follow the keyword argument input convention below:

\itemize{
\item{Each competing model is specified using keyword argument name starting with \emph{compModel} prefix.}
\item{Use \emph{.n} \emph{compModel} prefix in keyword argument name to represent applying to \emph{n}th intervention.}
}
The input to each outcome or competing model keyword argument is a model statement formula.
Please refer to the examples section for more examples.}
}
\value{
A list containing the following components:
\item{summary}{A summary table containing the estimated ICE risk, risk ratio, risk difference. If \code{bootstrap} is TRUE, then the table also includes standard error and confidence interval for ICE risk, risk ratio, and risk difference of each intervention.}
\item{risk.over.time}{A data frame containing the estimated ICE risk at each time point for each intervention.}
\item{models}{A list containing sublists whose names are the specified intervention descriptions, and each sublist contains the fitted models for the outcome, the treatment (if applicable), and the competing event (if applicable) of the corresponding intervention.}
\item{model.summary}{A list containing sublists whose names are the specified intervention descriptions, and each sublist contains the summary of the fitted models for the corresponding intervention.}
\item{model.stderr}{A list containing sublists whose names are the specified intervention descriptions, and each sublist contains the standard errors of the coefficients of the fitted models for the corresponding intervention.}
\item{model.vcov}{A list containing sublists whose names are the specified intervention descriptions, and each sublist contains the variance-covariance matrices of the parameters of the fitted models for the corresponding intervention.}
\item{model.rmse}{A list containing sublists whose names are the specified intervention descriptions, and each sublist contains root mean square error (RMSE) values of the fitted models for the corresponding intervention.}
\item{boot.data}{A list containing all the bootstrapped data if \code{boostrap} is set to \code{TRUE}.}
\item{boot.models}{A list containing the fitted models for the outcome, the treatment (if applicable), and the competing event (if applicable) on the bootstrapped samples.}
\item{boot.summary}{A list containing the summary of the fitted models on the bootstrapped samples.}
\item{boot.stderr}{A list containing the standard errors of the coefficients of the fitted models on the bootstrapped samples.}
\item{boot.vcov}{A list containing the variance-covariance matrices of the parameters of the fitted models on the bootstrapped samples.}
\item{boot.rmse}{A list containing the root mean square error (RMSE) values of the fitted models on the bootstrapped samples.}
\item{estimator.type}{A string specifying the type of ICE estimator.}
}
\description{
This function estimates the risk over time for survival outcome using the given observed data set following
multiple user-defined intervention strategies by the parametric g-formula iterative conditional expectation (ICE)
estimator. This function allows users to access all singly robust and doubly robust ICE estimators:
classical pooling over treatment history ICE estimator, classical stratifying on treatment history ICE estimator,
hazard-based pooling over treatment history ICE estimator, hazard-based stratifying on treatment history ICE estimator,
and a doubly robust inverse probability weighted ICE estimator. Please see Wen et al. (2021) more details regarding the parameter g-formula
iterative conditional expectation estimator.
}
\details{
Users could specify which version ICE estimator to use through \code{estimator}.
\itemize{
\item{\code{pool(hazard = F)} specifies the classical pooling over treatment history ICE estimator. }
\item{\code{pool(hazard = T)} specifies the hazard-based pooling over treatment history ICE estimator. }
\item{\code{strat(hazard = F)} specifies the classical stratifying on treatment history ICE estimator. }
\item{\code{strat(hazard = T)} specifies the hazard-based stratify on treatment history ICE estimator. }
\item{\code{weight(treat_model)} specifies the IP weighted ICE estimator
where \code{treat_model} specifies the treatment model.}
}
For stratified ICE, the estimator will automatically add in the treatment variables that are not intervened as covariates
into the model specification for each time point. To provide flexible choices on the model specification inputs, we also
provide keyword argument options for users to specify a different model statement, which is different from the model statement specified
in \code{outcome_model} or \code{competing_model}, for user-chosen interventions.
The input to each outcome or competing model keyword argument is a model statement formula.
To specify different outcome model for different intervention, please follow the keyword argument input convention below:

\itemize{
\item{Each outcome model is specified using keyword argument name starting with \emph{outcomeModel} prefix.}
\item{Use \emph{.n} \emph{outcomeModel} prefix in keyword argument name to specify which intervention being applied to,
where \emph{n} represents the \emph{n}th intervention.}
}

To specify different competing model for different intervention, please follow the keyword argument input convention below:

\itemize{
\item{Each competing model is specified using keyword argument name starting with \emph{compModel} prefix.}
\item{Use \emph{.n} \emph{compModel} prefix in keyword argument name to represent applying to \emph{n}th intervention.}
}

If an outcome or competing model is specified for an intervention using keyword argument, then the outcome or competing model will be used for the intervention.
If there is no outcome or competing model keyword argument specified, the outcome or competing model specified in \code{outcome_model} or \code{competing_model} will be
used for the intervention.

For example, the following input specifies \code{Y ~ L1} as outcome model for intervention 1,
\code{D ~ L1} as competing model for intervention 2,
\code{D ~ L1 + L2} as competing model for intervention 1,
\code{Y ~ L1 + L2} as outcome model for intervention 2.

\code{fit_hazard_strat <- ice(data = data, K = 4, id = "id", time_name = "t0", } \cr
\code{outcome_name = "Y", censor_name = "C", competing_name = "D", } \cr
\code{estimator = strat(hazard = T), comp_effect = 1, } \cr
\code{censor_model = C ~ L1 + L2, ref_idx = 0,} \cr
\code{int_descript = c("Static Intervention", "Dynamic Intervention"),} \cr
\code{outcome_model = Y ~ L1 + L2,} \cr
\code{competing_model = D ~ L1 + L2,} \cr
\code{intervention1.A1 = list(static(0)),} \cr
\code{intervention1.A2 = list(static(1)),} \cr
\code{intervention2.A1 = list(dynamic("absorbing", "L1", "=", 1)),} \cr
\code{intervention2.A2 = list(dynamic("compare", "L1", "=", 0)),} \cr
\code{outcomeModel.1 = Y ~ L1,} \cr
\code{compModel.2 = D ~ L1} \cr

Because the keyword argument for outcome model is not specified for intervention 2, the outcome model for intervention 2 is
\code{Y ~ L1 + L2} as specified in \code{outcome_model}.
Similarly, because the keyword argument for competing model is not specified for intervention 1, the competing model for intervention 1 is
\code{D ~ L1 + L2} as specified in \code{competing_model}.
Please see more examples in the examples section.
Note that for stratified ICE in the case of direct effect, the keyword argument competing model statement inputs are ignored since 
the competing model is used for nonparametric risk estimation.

Users could specify user-defined interventions or implement built-in interventions provided by the package
using the intervention input convention described in the parameter description section.

For the package built-in intervention strategy functions, if one wants to implement the strategy within the ICE method,
please follow the instructions described below. Furthermore, users could assess the intervened values by
specifying additional arguments following the detailed documentation for each function.
Built-in interventions include following:
\itemize{
\item{Always Treat:} {\code{static(1)}}
\item{Never Treat:} {\code{static(0)}}
\item{Dynamic Treat Based on \code{var}:} {\code{dynamic(type, var, direction, value)}. \code{type} could be "compare" or "absorbing" and \code{direction} could be \code{>, >=, <, <=, =, !=}.
If \code{type} is "compare", then treat if \code{var} > or >= or < or <= or = or != \code{value} and not treat otherwise. If \code{type} is "absorbing", then implements always treat if treatment is initiated
based on the description of the "compare" type.}
\item{Threshold Intervention Based on \code{var}:} {\code{threshold(lower_bound, upper_bound, var)}. If treatment value is within the range between \code{lower_bound} and \code{upper_bound} inclusively,
then follow the natural value of the treatment. Otherwise, if the treatment value is below \code{lower_bound}, then set to \code{lower_bound}.
If the treatment value is above the upper bound, then set to \code{upper_bound}.}
\item{Grace Period:} {\code{grace_period(type, nperiod, var, value)}. \code{type} could be "uniform" or "natural". \code{nperiod} specifies the length of grace period.
When a defined intervention condition based on \code{var} = \code{value} is met, initiate treatment in \code{nperiod} time units. If there is no intervention,
follow the observed treatment initiation distribution (for natural grace period) or the uniform distribution of treatment initiation (for uniform grace period). }
\item{User-Defined Intervention:} {The output of the user-defined intervention should be a vector of intervened value of the intervention variable for each individual at each time point in the size as the number of rows in \code{data}.}
}
Some examples are provided in the example section.

In order to obtain an inverse probability (IP) weighted natural course risk based on the observed data, users must specify
a censoring variable through \code{censor_name} and a corresponding censoring model through \code{censor_model}. Please see
Chiu et al. (2023) for more details regarding the IP weighted estimate of the natural course risk.

If competing event exists in the data, users need to specify the name of the competing variable through \code{competing_name} and
the model specification through \code{competing_model} for hazard-based ICE estimator. Users need to specify whether to treat
the competing event as censoring or total effect through \code{total_effect}.
}
\examples{

data <- gfoRmulaICE::data
dynamic_cat <- case_when(data$L2 < 0 ~ 1,
data$L2 >= 0 & data$L2 < 2 ~ 2,
T ~ 3)

# For the following examples, we consider two interventions.
# Intervention 1 is a static intervention, and
# intervention 2 is a dynamic intervention, where the intervention
# for A1 is user-defined.

# Hazard based stratified ICE,
# competing event as total effect for all interventions,
# with Y ~ L1 + L2 as outcome model for both intervention 1 and intervention 2,
# and D ~ L1 + L2 as competing model for both intervention 1 and intervention 2,
# bootstrap with 1000 samples with normal quantile,
# natural course as the reference intervention

ice_strat_haz <- ice(data = data, time_points = 4, id = "id", time_name = "t0",
censor_name = "C", outcome_name = "Y",
compevent_name = "D",
outcome_model = Y ~ L1 + L2, censor_model = C ~ L1 + L2,
competing_model = D ~ L1 + L2,
comp_effect = 1,
ref_idx = 0,
estimator = strat(hazard = T),
nsamples = 1000, ci_method = "normal",
int_descript = c("Static Intervention",
"Dynamic Intervention"),
intervention1.A1 = list(static(3)),
intervention1.A2 = list(static(1)),
intervention2.A1 = list(dynamic_cat),
intervention2.A2 = list(dynamic("compare", "L1", "=", 0))
)

plot_risk(ice_strat_haz)

# Hazard based stratified ICE,
# competing event as total effect for all interventions,
# with Y ~ L1 + L2 as outcome model for intervention 1,
# Y ~ L1 as outcome model for intervention 2,
# D ~ L1 as competing model for intervention 1,
# and D ~ L1 + L2 as competing model for intervention 2,
# bootstrap with 1000 samples with normal quantile,
# natural course as the reference intervention

ice_strat_haz <- ice(data = data, time_points = 4, id = "id", time_name = "t0",
censor_name = "C", outcome_name = "Y",
compevent_name = "D",
outcome_model = Y ~ L1, censor_model = C ~ L1,
competing_model = D ~ L1,
comp_effect = 1,
ref_idx = 0,
estimator = strat(hazard = T),
nsamples = 1000, ci_method = "normal",
int_descript = c("Static Intervention",
"Dynamic Intervention"),
intervention1.A1 = list(static(3)),
intervention1.A2 = list(static(1)),
intervention2.A1 = list(dynamic_cat),
intervention2.A2 = list(dynamic("compare", "L1", "=", 0)),
outcomeModel.1 = Y ~ L1 + L2,
compModel.2 = D ~ L1 + L2
)

plot_risk(ice_strat_haz)

# Classical pooled ICE,
# competing event as direct effect for all interventions,
# bootstrap with 1000 samples using empirical quantile,
# natural course as the reference intervention.

ice_pool_classic <- ice(data = data, time_points = 4, id = "id", time_name = "t0",
censor_name = "C", outcome_name = "Y",
compevent_name = "D",
comp_effect = 0,
outcome_model = Y ~ L1 + A1 + A2, censor_model = C ~ L1 + A1 + A2,
competing_model = D ~ L1 + A1 + A2,
ref_idx = 0,
estimator = pool(hazard = F),
nsamples = 1000, ci_method = "percentile",
int_descript = c("Static Intervention",
"Dynamic Intervention"),
intervention1.A1 = list(static(3)),
intervention1.A2 = list(static(1)),
intervention2.A1 = list(dynamic_cat),
intervention2.A2 = list(dynamic("compare", "L1", "=", 0))
)

plot_risk(ice_pool_classic)

# Hazard based pooled ICE,
# competing event as direct effect for all interventions,
# bootstrap with 1000 samples using empirical quantile,
# always treat as the reference intervention.

ice_pool_haz <- ice(data = data, time_points = 4, id = "id", time_name = "t0",
censor_name = "C", outcome_name = "Y",
compevent_name = "D",
comp_effect = 0,
outcome_model = Y ~ L1 + A1 + A2, censor_model = C ~ L1 + A1 + A2,
competing_model = D ~ L1 + A1 + A2,
ref_idx = 1,
estimator = pool(hazard = T),
nsamples = 1000, ci_method = "percentile",
int_descript = c("Static Intervention",
"Dynamic Intervention"),
intervention1.A1 = list(static(3)),
intervention1.A2 = list(static(1)),
intervention2.A1 = list(dynamic_cat),
intervention2.A2 = list(dynamic("compare", "L1", "=", 0))
)

plot_risk(ice_pool_haz)

# Doubly robust ICE,
# competing event as direct effect for all interventions,
# with Y ~ L1 as outcome model for both intervention 1 and intervention 2,
# and D ~ L1 as competing model for both intervention 1 and intervention 2,
# bootstrap with 1000 samples using normal quantile,
# natural course as the reference intervention.

ice_weight <- ice(data = data, time_points = 4, id = "id", time_name = "t0",
censor_name = "C", outcome_name = "Y",
compevent_name = "D",
comp_effect = 0,
outcome_model = Y ~ L1, censor_model = C ~ L1,
competing_model = D ~ L1,
ref_idx = 0,
estimator = weight(list(A1 ~ L1 + L2, A2 ~ L1 + L2)),
nsamples = 1000, ci_method = "normal",
int_descript = c("Static Intervention",
"Dynamic Intervention"),
intervention1.A1 = list(static(3)),
intervention1.A2 = list(static(1)),
intervention2.A1 = list(dynamic_cat),
intervention2.A2 = list(dynamic("compare", "L1", "=", 0))
)

plot_risk(ice_weight)

# Doubly robust ICE,
# competing event as direct effect for all interventions,
# with Y ~ L1 as outcome model for intervention 1,
# Y ~ L1 + L2 as outcome model for intervention 2,
# D ~ L1 + L2 as competing model for intervention 1,
# D ~ L1 as competing model for intervention 2,
# bootstrap with 1000 samples using normal quantile,
# natural course as the reference intervention.

ice_weight <- ice(data = data, time_points = 4, id = "id", time_name = "t0",
censor_name = "C", outcome_name = "Y",
compevent_name = "D",
comp_effect = 0,
outcome_model = Y ~ L1, censor_model = C ~ L1,
competing_model = D ~ L1,
ref_idx = 0,
estimator = weight(list(A1 ~ L1 + L2, A2 ~ L1 + L2)),
nsamples = 1000, ci_method = "normal",
int_descript = c("Static Intervention",
"Dynamic Intervention"),
intervention1.A1 = list(static(3)),
intervention1.A2 = list(static(1)),
intervention2.A1 = list(dynamic_cat),
intervention2.A2 = list(dynamic("compare", "L1", "=", 0)),
outcomeModel.2 = Y ~ L1 + L2,
compModel.1 = D ~ L1 + L2
)

plot_risk(ice_weight)


# The following example implements the natural value intervention on L1
# (i.e. if L1 > 0, then replace the observed value of L1 by 0,
# and keep the observed value of L1 otherwise),
# dynamic intervention on L1 (treat when L1 = 0)
# with uniform grace period of 2 periods, and
# threshold intervention when the natural value of L2 at time t is lower
# than -3, set its value to -3. Otherwise, do not intervene.

# Classical pooled ICE,
# competing event as direct effect for all interventions,
# bootstrap with 1000 samples using empirical quantile,
# natural course as the reference intervention.

ice_pool_grace_period <- ice(data = data, time_points = 4, id = "id", time_name = "t0",
censor_name = "C", outcome_name = "Y",
compevent_name = "D",
comp_effect = 0,
outcome_model = Y ~ L1 + L2 + A1 + A2, censor_model = C ~ L1 + L2 + A1 + A2,
competing_model = D ~ L1 + L2 + A1 + A2,
ref_idx = 0,
estimator = pool(hazard = F),
nsamples = 1000, ci_method = "percentile",
int_descript = c("Grace Period", "Threshold Intervention")
intervention1.A2 = list(grace_period("uniform", 2, "L1", 0)),
intervention2.L2 = list(threshold(-3))
)

plot_risk(ice_pool_grace_period)


}
\references{
Wen L, Young JG, Robins JM, Hernán MA. Parametric g-formula implementations for causal survival analyses. Biometrics. 2021;77(2):740-753.

McGrath S, Lin V, Zhang Z, Petito LC, Logan RW, Hernán MA, and JG Young. gfoRmula: An R package for estimating the effects of sustained treatment strategies via the parametric g-formula. Patterns. 2020;1:100008.

Young JG, Herńan MA, Robins JM. Identification, estimation and approximation of risk under interventions that depend on the natural value of treatment using observational data. Epidemiologic Methods. 2014;3(1):1-19.

Young JG, Vatsa R, Murray EJ, Hernán MA. Interval-cohort designs and bias in the estimation of per-protocol effects: a simulation study. Trials. 2019;20(1):552.

Díaz, I, Williams, N, Hoffman, KL, & Schenck, EJ. Nonparametric causal effects based on longitudinal modified treatment policies. Journal of the American Statistical Association. 2021;118(542), 846–857.

Young JG, Stensrud MJ, Tchetgen Tchetgen EJ, Hernán MA. A causal framework for classical statistical estimands in failure-time settings with competing events. Statistics in medicine. 2020;39(8):1199-1236.

Wen L, Hernán MA, Robins JM. Multiply robust estimators of causal effects for survival outcomes. Scandinavian journal of statistics, theory and applications. 2022;49(3):1304-1328.

Haneuse S, Rotnitzky A. Estimation of the effect of interventions that modify the received treatment. Statistics in medicine. 2013;32(30):5260-5277.

McGrath S, Young JG, Hernán MA. Revisiting the g-null Paradox. Epidemiology. 2022;33(1):114-120.

Chiu YH, Wen L, McGrath S, Logan R, Dahabreh IJ, Hernán MA. Evaluating model specification when using the parametric g-formula in the presence of censoring. American Journal of Epidemiology. 2023;192:1887–1895.
}
