% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{natural_course}
\alias{natural_course}
\title{Natural Course}
\usage{
natural_course(data, treat_var)
}
\arguments{
\item{data}{a data frame containing the observed data.}

\item{treat_var}{a string specifying the treatment variable in \code{data}.}
}
\value{
a vector containing the intervened values of the same size as the number of rows in \code{data}.
}
\description{
This function specifies the natural course intervention on the treatment variable in \code{data}.
}
\examples{
data <- gfoRmulaICE::compData
natural_course <- natural_course(data = data, treat_var = "A1")

}
