% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_risk}
\alias{plot_risk}
\title{Plot risk estimated by ICE estimator over time}
\usage{
plot_risk(..., plot_np = T, label = 0)
}
\arguments{
\item{...}{ICE estimator objects.}

\item{label}{a numeric specifying which time label is used in x axis. 0 represents using generic numerical time index, and
1 represents using the original time index in the data set. Default is 0.}

\item{plot_obs}{a logical indicating whether to add the observed risk over time to the plot.
Default is TRUE}
}
\value{
a plot for risks of all the interventions specified in \code{...}.
}
\description{
This function provides visualization of estimated risk on all user-defined interventions
from the fitted ICE estimator object with natural course risk over time.
(This function is going to be converted to the S3 method in R so it is named "plot_risk" for now.
After conversion, users could use the base function plot().)
}
\examples{

ice_fit1 <- ice(
data = data, 
time_points = 4, 
id = "id", 
time_name = "t0",
censor_name = "C", 
outcome_name = "Y",
compevent_name = "D",
comp_effect = 0,
outcome_model = Y ~ L1 + L2 + A1 + A2, 
censor_model = C ~ L1 + L2 + A1 + A2,
competing_model = D ~ L1 + L2 + A1 + A2,
ref_idx = 0,
estimator = pool(hazard = F),
int_descript = "Static Intervention",
intervention1.A1 = list(static(3)),
intervention1.A2 = list(static(1))
)

ice_fit2 <- ice(
data = data, 
time_points = 4, 
id = "id", 
time_name = "t0",
censor_name = "C", 
outcome_name = "Y",
compevent_name = "D",
comp_effect = 0,
outcome_model = Y ~ L1 + L2 + A1 + A2, 
censor_model = C ~ L1 + L2 + A1 + A2,
competing_model = D ~ L1 + L2 + A1 + A2,
ref_idx = 0,
estimator = pool(hazard = T),
int_descript = "Static Intervention",
intervention1.A1 = list(static(3)),
intervention1.A2 = list(static(1))
)

plot_risk(ice_fit1, ice_fit2)

}
