% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_risk}
\alias{plot_risk}
\title{Plot for ICE estimator objects}
\usage{
plot_risk(..., plot_np = T, label = 0)
}
\arguments{
\item{...}{ICE estimator objects.}

\item{label}{a number specifying which time label is used in x-axis. 0 represents using generic numerical time index, and
1 represents using the original time label from the data. Default is 0.}

\item{plot_obs}{a logical value indicating whether to plot the observed risk over time.
Default is \code{TRUE}.}
}
\value{
a plot for risks of all the interventions specified in \code{...}.
}
\description{
This function provides visualization of estimated risk 
for all specified interventions, estimated natural course risk, and observed risk at each time point.
(This function is going to be converted to the S3 method in R so it is named "plot_risk" for now.
After conversion, users could use the base function plot().)
}
\examples{

ice_fit1 <- ice(
data = data, 
time_points = 4, 
id = "id", 
time_name = "t0",
censor_name = "C", 
outcome_name = "Y",
compevent_name = "D",
comp_effect = 0,
outcome_model = Y ~ L1 + L2 + A1 + A2, 
censor_model = C ~ L1 + L2 + A1 + A2,
ref_idx = 0,
estimator = pool(hazard = F),
int_descript = "Static Intervention",
intervention1.A1 = list(static(3)),
intervention1.A2 = list(static(1))
)

ice_fit2 <- ice(
data = data, 
time_points = 4, 
id = "id", 
time_name = "t0",
censor_name = "C", 
outcome_name = "Y",
compevent_name = "D",
comp_effect = 0,
outcome_model = Y ~ L1 + L2 + A1 + A2, 
censor_model = C ~ L1 + L2 + A1 + A2,
competing_model = D ~ L1 + L2 + A1 + A2,
ref_idx = 0,
estimator = pool(hazard = T),
int_descript = "Static Intervention",
intervention1.A1 = list(static(3)),
intervention1.A2 = list(static(1))
)

plot_risk(ice_fit1, ice_fit2)

}
