% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{pool}
\alias{pool}
\title{Indicator for the pooling over treatment history ICE estimator}
\usage{
pool(hazard)
}
\arguments{
\item{hazard}{a logical value indicating whether to use hazard-based ICE estimator.}
}
\value{
a logical value on whether to use hazard-based ICE estimator.
}
\description{
This function identifies the pooling over treatment history ICE estimator. The classical pooling over
treatment history ICE estimator is specified by \code{pool(hazard = FALSE)}. The hazard based pooling over treatment
history ICE estimator is specified by \code{pool(hazard = TRUE)}.
}
