% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{static}
\alias{static}
\title{Static Intervention Function}
\usage{
static(value, data = interv_data)
}
\arguments{
\item{value}{a numeric specifying the intervention value. 1 for always treat and 0 for never treat.}

\item{data}{a data frame containing the observed data.}
}
\value{
a vector containing the intervened value in the same size as the number of rows in \code{data}.
}
\description{
This function implements the static intervention, specifically always treat and never treat.
Always treat refers to the constant treatment across all time points. Never treat refers to no treatment across all time points.
}
\examples{
data <- readRDS("test_data_competing.rds")
always_treat <- static(value = 1, data = data)
always_treat
}
