% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{strat}
\alias{strat}
\title{Indicator for the stratifying on treatment history ICE estimator}
\usage{
strat(hazard)
}
\arguments{
\item{hazard}{a logical value indicating whether to use hazard-based ICE estimator.}
}
\value{
a logical value on whether to use hazard-based ICE estimator.
}
\description{
This function identifies the stratifying on treatment history ICE estimator. The classical stratifying on 
treatment history ICE estimator is specified by \code{strat(hazard = FALSE)}. The hazard based stratifying on treatment
history ICE estimator is specified by \code{strat(hazard = TRUE)}.
}
