% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{weight}
\alias{weight}
\title{Indicator for the weighted ICE estimator in the main function \code{ice}}
\usage{
weight(treat_model = list())
}
\arguments{
\item{treat_model}{a list of formulas specifying the model statement for the corresponding treatment variable used in the doubly robust ICE estimator. The length of list must match with
the number of treatment variables.}
}
\value{
the model specification for each treatment model and
the observed treatment variable names extracted from the specified model.
}
\description{
This function identifies the doubly robust weighted ICE estimator. The treatment models could be specified by
\code{treat_model} and the treatment variables could be specified by \code{obs_treatment_varnames}.
}
